moving lockfeed extraction from dashboard to creatorLocks component
