+++ unlock-app/src/components/content/DashboardContent.js
@@ -16,11 +16,9 @@ import {
 } from '../interface/buttons/ActionButton'
 import { showForm, hideForm } from '../../actions/lockFormVisibility'
 
-// TODO : move lockFeed extraction in CreatorLocks since it's just being passed down there
 export const DashboardContent = ({
   account,
   network,
-  lockFeed,
   formIsVisible,
   showForm,
   hideForm,
@@ -42,7 +40,7 @@ export const DashboardContent = ({
                 Create Lock
               </CreateLockButton>
             </AccountWrapper>
-            <CreatorLocks lockFeed={lockFeed} />
+            <CreatorLocks />
             <DeveloperOverlay />
           </BrowserOnly>
         )}
@@ -54,53 +52,23 @@ export const DashboardContent = ({
 DashboardContent.propTypes = {
   account: UnlockPropTypes.account,
   network: UnlockPropTypes.network.isRequired,
-  lockFeed: PropTypes.arrayOf(UnlockPropTypes.lock),
   formIsVisible: PropTypes.bool.isRequired,
   showForm: PropTypes.func.isRequired,
   hideForm: PropTypes.func.isRequired,
 }
 
 DashboardContent.defaultProps = {
-  lockFeed: [],
   account: null,
 }
 
 export const mapStateToProps = ({
-  transactions,
-  locks,
   account,
   network,
   lockFormStatus: { visible },
 }) => {
-  // We want to display newer locks first, so sort the locks by blockNumber in descending order
-  const locksComparator = (a, b) => {
-    // Newly created locks may not have a transaction associated just yet
-    // -- those always go right to the top
-    if (!transactions[a.transaction]) {
-      return -1
-    }
-    if (!transactions[b.transaction]) {
-      return 1
-    }
-    return (
-      transactions[b.transaction].blockNumber -
-      transactions[a.transaction].blockNumber
-    )
-  }
-
-  // Only show the current account's locks
-  const locksFilter = lock => {
-    return lock.owner === account.address
-  }
-
-  const lockFeed = Object.values(locks)
-    .filter(locksFilter)
-    .sort(locksComparator)
-
   return {
     account: account,
     network: network,
-    lockFeed,
     formIsVisible: visible,
   }
 }
