+++ unlock-app/src/components/creator/CreatorLocks.js
@@ -53,9 +53,9 @@ export const CreatorLocks = props => {
 }
 
 CreatorLocks.propTypes = {
+  lockFeed: PropTypes.arrayOf(UnlockPropTypes.lock),
   createLock: PropTypes.func.isRequired,
   formIsVisible: PropTypes.bool.isRequired,
-  lockFeed: PropTypes.arrayOf(UnlockPropTypes.lock),
   loading: PropTypes.bool,
   hideForm: PropTypes.func.isRequired,
 }
@@ -70,8 +70,40 @@ const mapDispatchToProps = dispatch => ({
   hideForm: () => dispatch(hideForm()),
 })
 
-export const mapStateToProps = ({ loading, lockFormStatus: { visible } }) => {
+export const mapStateToProps = ({
+  account,
+  loading,
+  lockFormStatus: { visible },
+  transactions,
+  locks,
+}) => {
+  // We want to display newer locks first, so sort the locks by blockNumber in descending order
+  const locksComparator = (a, b) => {
+    // Newly created locks may not have a transaction associated just yet
+    // -- those always go right to the top
+    if (!transactions[a.transaction]) {
+      return -1
+    }
+    if (!transactions[b.transaction]) {
+      return 1
+    }
+    return (
+      transactions[b.transaction].blockNumber -
+      transactions[a.transaction].blockNumber
+    )
+  }
+
+  // Only show the current account's locks
+  const locksFilter = lock => {
+    return lock.owner === account.address
+  }
+
+  const lockFeed = Object.values(locks)
+    .filter(locksFilter)
+    .sort(locksComparator)
+
   return {
+    lockFeed,
     loading: !!loading,
     formIsVisible: visible,
   }
